Deleting link after 6 hours

+++ app/jobs/pending_invitations_cleanup_job.rb
@@ -2,6 +2,8 @@ class PendingInvitationsCleanupJob < ApplicationJob
   queue_as :default
 
   def perform(*args)
-    Invitation.expired.delete_all
+    if Invitation.username.nil?
+      Invitation.expired.delete_all
+    end
   end
 end
